*** empty log message ***

+++ magick/animate.c
@@ -640,7 +640,7 @@ MagickExport void XAnimateBackgroundImage(Display *display,
   */
   resources.colormap=SharedColormap;
   display_image=image_list[0];
-  for (scene=0; scene < (int) number_scenes; scene++)
+  for (scene=0; scene < (long) number_scenes; scene++)
   {
     if ((resource_info->map_type != (char *) NULL) ||
         (visual_info->storage_class == TrueColor) ||
@@ -783,7 +783,7 @@ MagickExport void XAnimateBackgroundImage(Display *display,
   window_info.matte_pixmaps[0]=window_info.pixmap;
   scene_info.pixels=(unsigned long *) NULL;
   scene_info.gamma_map=(XColor *) NULL;
-  for (scene=1; scene < (int) number_scenes; scene++)
+  for (scene=1; scene < (long) number_scenes; scene++)
   {
     /*
       Create X image.
@@ -851,7 +851,7 @@ MagickExport void XAnimateBackgroundImage(Display *display,
   event.type=Expose;
   do
   {
-    for (scene=0; scene < (int) number_scenes; scene++)
+    for (scene=0; scene < (long) number_scenes; scene++)
     {
       if (XEventsQueued(display,QueuedAfterFlush) > 0)
         {
@@ -1247,10 +1247,10 @@ MagickExport Image *XAnimateImages(Display *display,
   if (image_list == (Image **) NULL)
     MagickError(ResourceLimitError,"Unable to animate images",
       "Memory allocation failed");
-  for (scene=0; scene < (int) number_scenes; scene++)
+  for (scene=0; scene < (long) number_scenes; scene++)
     if (image_list[scene]->scene == 0)
       break;
-  if (scene == (int) number_scenes)
+  if (scene == (long) number_scenes)
     qsort((void *) image_list,number_scenes,sizeof(Image *),SceneCompare);
   /*
     Initialize Standard Colormap.
@@ -1258,7 +1258,7 @@ MagickExport Image *XAnimateImages(Display *display,
   nexus=(Image *) NULL;
   display_image=image_list[0];
   (void) TransformRGBImage(display_image,RGBColorspace);
-  for (scene=0; scene < (int) number_scenes; scene++)
+  for (scene=0; scene < (long) number_scenes; scene++)
   {
     if ((resource_info->map_type != (char *) NULL) ||
         (visual_info->storage_class == TrueColor) ||
@@ -1636,7 +1636,7 @@ MagickExport Image *XAnimateImages(Display *display,
   windows->image.matte_pixmaps[0]=windows->image.matte_pixmap;
   scene_info.pixels=(unsigned long *) NULL;
   scene_info.gamma_map=(XColor *) NULL;
-  for (scene=1; scene < (int) number_scenes; scene++)
+  for (scene=1; scene < (long) number_scenes; scene++)
   {
     /*
       Create X image.
